\name{dental}
\alias{dental}
\title{
A Data Set of Orthodontic Measurements on Children
}
\description{
The dental data consists of 27 subjects from the University of North Carolina Dental School.
}
\usage{
dental
}
%- maybe also 'usage' for other objects documented here.
\format{
 \itemize{
 \item{subject:}{cluster id}
  \item{gender:}{the sex of subject}
  \item{age_8:}{the measurements at age 8}
  \item{age_10:}{the measurements at age 10}
  \item{age_12:}{the measurements at age 12}
  \item{age_14:}{the measurements at age 14}
 }
}  
\details{
The data was collected by investigators at the University of North Carolina Dental School. It is a study of orthodontic measurements on children including 11 girls and 16 boys measured at ages 8, 10, 12, and 14.
}

\source{
Pottho R. F. and Roy, S. W. A generalized multivariate analysis of variance model
useful especially for growth curve problems. Biometrika 1964;51:313-326
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{cluster}
\keyword{GEE}% __ONLY ONE__ keyword per line
