% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEE.var.md_bin.aggr.R
\name{GEE.var.md_bin.aggr}
\alias{GEE.var.md_bin.aggr}
\title{GEE.var.md_bin.aggr}
\usage{
GEE.var.md_bin.aggr(formula, id, data, corstr = "independence")
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): 
a formula expression as for other regression models to be fitted, 
of the form c(successes,failures) ~ predictors. The details of formula specification can be seen in glm() and gee().}

\item{id}{a vector which identifies the clusters. The length of id should be the same as the total number of observations. 
Data is assumed to be sorted so that observations on a cluster are contiguous rows for all entities in the formula.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) 
containing the variables in the model. If not found in data, the variables are taken from environment(formula), 
typically the environment from which glm is called.}

\item{corstr}{a character string specifying working correlation structure: 
"independence", "AR-M","exchangeable", "unstructured"  are possible.}
}
\description{
GEE.var.md_bin.aggr
}
