\name{toenail}  
\alias{toenail}
\title{
Toenail infection data from a multicenter study
}
\description{
The data frame consists of 250 patients from a multicenter study comparing to treatments for toenail infection.
}
\usage{
toenail
}
%- maybe also 'usage' for other objects documented here.
\format{
\itemize{
  \item ID: patient id. The variable for cluster id
  \item Response: toenail infection status: 0=none or mild; 1=moderate or severe
  \item Treatment: 1=oral treatment A; 0=oral treatment B
  \item Month: the exact time point of measurement in month
  \item Visit: the visit index
 }
}
\details{
In a randomized, double-blinded multicenter study, treatment A and B were compared to treat toenail infection disease. Patientes were evaluated for the degree of onycholysis. There are 1908 measurements in total.  
}

\source{
De Backer M, De Vroey C, Lesaffre E, Scheys I, De Keyser P. Twelve weeks of continuous oral therapy for toenail onychomycosis caused by dermatophytes: a double-blind comparative trial of terbinafine 250 mg/day versus itraconazole 200 mg/day. Journal of the American Academy of Dermatology 1998; 38: 57-63.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{toenail}
\keyword{GEE}% __ONLY ONE__ keyword per line
